requirements_.txt
crewai
crewai-tools
langchain
langchain-community
openai
python-dotenv
faiss-cpu
langchain-openai
langchain-text-splitters
pandas